Nuspec update
