+++ IntrinioRealTimeClient.nuspec
@@ -2,17 +2,17 @@
 <package>
 	<metadata>
 		<id>IntrinioRealTimeClient</id>
-		<version>10.0.0</version>
-		<title>Intrinio SDK for Real-Time Stock Prices</title>
+		<version>11.0.0</version>
+		<title>Intrinio SDK for Real-Time Stock and Option Prices</title>
 		<authors>Intrinio</authors>
 		<owners>Intrinio</owners>
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<licenseUrl>https://licenses.nuget.org/MIT</licenseUrl>
 		<projectUrl>https://github.com/intrinio/intrinio-realtime-csharp-sdk</projectUrl>
-		<description>Intrinio provides real-time stock prices via a two-way WebSocket connection.</description>
-		<releaseNotes>Version 10.0.0 release.</releaseNotes>
+		<description>Intrinio provides real-time stock and option prices via a two-way WebSocket connection.</description>
+		<releaseNotes>Version 11.0.0 release.</releaseNotes>
 		<copyright>Copyright 2024 Intrinio</copyright>
-		<tags>fintech stocks prices websocket real-time market finance</tags>
+		<tags>fintech stocks options prices websocket real-time market finance</tags>
 		<dependencies>
 			<group targetFramework="net8.0">
 				<dependency id="FSharp.Core" version="7.0.400" />
@@ -30,14 +30,14 @@
 		</dependencies>
 	</metadata>
 	<files>
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealtimeMultiExchange.dll" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealtimeMultiExchange.xml" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealtimeMultiExchange.pdb" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/config.json" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealTimeSDK.deps.json" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealTimeSDK.dll" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealTimeSDK.pdb" target="lib/net8.0" />
-		<file src="IntrinioRealTimeSDK/bin/Release/net8.0/IntrinioRealTimeSDK.runtimeconfig.json" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/IntrinioRealtime.dll" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/IntrinioRealtime.xml" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/IntrinioRealtime.pdb" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/config.json" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/SampleApp.deps.json" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/SampleApp.dll" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/SampleApp.pdb" target="lib/net8.0" />
+		<file src="SampleApp/bin/Release/net8.0/SampleApp.runtimeconfig.json" target="lib/net8.0" />
 		<file src="intrinio.png" target="intrinio.png" />
 		<file src="README.md" target="README.md" />
 	</files>
