+++ IntrinioRealTimeSDK.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "IntrinioRealtimeMultiExchange", "IntrinioRealtimeMultiExchange\IntrinioRealtimeMultiExchange.fsproj", "{3409E8EE-BB04-4E36-B2F8-029134727FFB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntrinioRealTimeSDK", "IntrinioRealTimeSDK\IntrinioRealTimeSDK.csproj", "{6E5BD6AB-9909-4980-BA1C-B0466445CD2B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntrinioRealtime", "IntrinioRealtime\IntrinioRealtime.csproj", "{BA956D20-BB6A-40B1-B204-46DDE5868FB9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleApp", "SampleApp\SampleApp.csproj", "{A013870E-4C79-4113-B7B2-12A2A864C2D7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +19,14 @@ Global
 		{3409E8EE-BB04-4E36-B2F8-029134727FFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3409E8EE-BB04-4E36-B2F8-029134727FFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3409E8EE-BB04-4E36-B2F8-029134727FFB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E5BD6AB-9909-4980-BA1C-B0466445CD2B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E5BD6AB-9909-4980-BA1C-B0466445CD2B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E5BD6AB-9909-4980-BA1C-B0466445CD2B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E5BD6AB-9909-4980-BA1C-B0466445CD2B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BA956D20-BB6A-40B1-B204-46DDE5868FB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BA956D20-BB6A-40B1-B204-46DDE5868FB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BA956D20-BB6A-40B1-B204-46DDE5868FB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BA956D20-BB6A-40B1-B204-46DDE5868FB9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A013870E-4C79-4113-B7B2-12A2A864C2D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A013870E-4C79-4113-B7B2-12A2A864C2D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A013870E-4C79-4113-B7B2-12A2A864C2D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A013870E-4C79-4113-B7B2-12A2A864C2D7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
