+++ IntrinioRealtime/IntrinioRealtime.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <Folder Include="Composite\" />
+      <Folder Include="Equities\" />
+      <Folder Include="Options\" />
+    </ItemGroup>
+
+</Project>
