+++ SampleApp/SampleApp.csproj
@@ -1,0 +1,25 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <PublishAot>true</PublishAot>
+        <InvariantGlobalization>true</InvariantGlobalization>
+        <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <ServerGarbageCollection>true</ServerGarbageCollection>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <Content Include="runtimeconfig.template.json" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\IntrinioRealtimeMultiExchange\IntrinioRealtimeMultiExchange.fsproj" />
+    </ItemGroup>
+
+</Project>
