+++ SampleApp/SampleApp.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <doc>
     <assembly>
-        <name>IntrinioRealTimeSDK</name>
+        <name>SampleApp</name>
     </assembly>
     <members>
     </members>
